chore(4484): use only one Team Contact component for both clouds
