+++ app/app/private-cloud/requests/(request)/[id]/request/page.tsx
@@ -16,11 +16,11 @@ import { z } from 'zod';
 import CancelRequest from '@/components/buttons/CancelButton';
 import PreviousButton from '@/components/buttons/Previous';
 import ProjectDescription from '@/components/form/ProjectDescriptionPrivate';
+import TeamContacts from '@/components/form/TeamContacts';
 import PageAccordion from '@/components/generic/accordion/PageAccordion';
 import AdditionalTeamMembers from '@/components/private-cloud/sections/AdditionalTeamMembers';
-import ModifiedTeamContacts from '@/components/private-cloud/sections/ModifiedTeamContacts';
 import Quotas from '@/components/private-cloud/sections/Quotas';
-import { GlobalRole } from '@/constants';
+import { GlobalRole, userAttributes } from '@/constants';
 import createClientPage from '@/core/client-page';
 import { usePrivateProductState } from '@/states/global';
 
@@ -82,8 +82,8 @@ export default privateCloudRequestRequest(({}) => {
       LeftIcon: IconUsersGroup,
       label: 'Team contacts',
       description: '',
-      Component: ModifiedTeamContacts,
-      componentArgs: { disabled: isDisabled, secondTechLead, secondTechLeadOnClick },
+      Component: TeamContacts,
+      componentArgs: { userAttributes, disabled: isDisabled, secondTechLead, secondTechLeadOnClick },
     },
     {
       LeftIcon: IconUsersGroup,
