+++ app/app/public-cloud/products/(product)/[licencePlate]/edit/page.tsx
@@ -19,12 +19,12 @@ import AccountCoding from '@/components/form/AccountCoding';
 import AccountEnvironmentsPublic from '@/components/form/AccountEnvironmentsPublic';
 import Budget from '@/components/form/Budget';
 import ProjectDescriptionPublic from '@/components/form/ProjectDescriptionPublic';
+import TeamContacts from '@/components/form/TeamContacts';
 import PageAccordion from '@/components/generic/accordion/PageAccordion';
 import FormErrorNotification from '@/components/generic/FormErrorNotification';
 import { openPublicCloudProductEditSubmitModal } from '@/components/modal/publicCloudProductEditSubmit';
 import AdditionalTeamMembers from '@/components/public-cloud/sections/AdditionalTeamMembers';
-import ModifiedTeamContacts from '@/components/public-cloud/sections/ModifiedTeamContacts';
-import { AGMinistries, GlobalRole } from '@/constants';
+import { AGMinistries, GlobalRole, userAttributes } from '@/constants';
 import createClientPage from '@/core/client-page';
 import { usePublicProductState } from '@/states/global';
 import { publicCloudEditRequestBodySchema } from '@/validation-schemas/public-cloud';
@@ -99,8 +99,8 @@ export default publicCloudProductEdit(({}) => {
       LeftIcon: IconUsersGroup,
       label: 'Team contacts',
       description: '',
-      Component: ModifiedTeamContacts,
-      componentArgs: { disabled: isDisabled, secondTechLead, secondTechLeadOnClick },
+      Component: TeamContacts,
+      componentArgs: { disabled: isDisabled, hasEA: true, userAttributes, secondTechLead, secondTechLeadOnClick },
     },
     {
       LeftIcon: IconUsersGroup,
