+++ app/app/public-cloud/products/(product)/create/page.tsx
@@ -17,11 +17,11 @@ import AccountCoding from '@/components/form/AccountCoding';
 import AccountEnvironmentsPublic from '@/components/form/AccountEnvironmentsPublic';
 import Budget from '@/components/form/Budget';
 import ProjectDescriptionPublic from '@/components/form/ProjectDescriptionPublic';
+import TeamContacts from '@/components/form/TeamContacts';
 import PageAccordion from '@/components/generic/accordion/PageAccordion';
 import FormErrorNotification from '@/components/generic/FormErrorNotification';
 import { openPublicCloudProductCreateSubmitModal } from '@/components/modal/publicCloudProductCreateSubmit';
-import ModifiedTeamContacts from '@/components/public-cloud/sections/ModifiedTeamContacts';
-import { GlobalRole } from '@/constants';
+import { GlobalRole, userAttributes } from '@/constants';
 import createClientPage from '@/core/client-page';
 import { existBilling } from '@/services/backend/billing';
 import { publicCloudCreateRequestBodySchema } from '@/validation-schemas/public-cloud';
@@ -87,8 +87,8 @@ export default publicCloudProductNew(({}) => {
       LeftIcon: IconUsersGroup,
       label: 'Team contacts',
       description: '',
-      Component: ModifiedTeamContacts,
-      componentArgs: { secondTechLead, secondTechLeadOnClick },
+      Component: TeamContacts,
+      componentArgs: { hasEA: true, userAttributes, secondTechLead, secondTechLeadOnClick },
     },
     {
       LeftIcon: IconMoneybag,
