+++ app/components/form/TeamContacts.tsx
@@ -1,136 +1,173 @@
+import { Avatar, Badge, Table, UnstyledButton, Button, Text, Group } from '@mantine/core';
+import { User } from '@prisma/client';
+import { IconMinus } from '@tabler/icons-react';
 import React from 'react';
 import { useFormContext } from 'react-hook-form';
-import { IMaskInput } from 'react-imask';
-import SecondTechLeadButton from '@/components/buttons/SecondTechLeadButton';
-import AsyncAutocomplete from '@/components/form/AsyncAutocomplete';
 import { openConfirmModal } from '@/components/modal/confirm';
-import { cn } from '@/utils/js';
+import { openUserPickerModal } from '@/components/modal/userPicker';
+import { formatFullName } from '@/helpers/user';
+import { getUserImageData } from '@/helpers/user-image';
+import { cn, formatDate } from '@/utils/js';
+import SupportContact from './SupportContact';
 
-const getErrorMessage = (error: any): string | undefined => {
-  if (error?.message) {
-    return error.message;
-  }
-  return undefined;
-};
+interface UserAttribute {
+  role: string;
+  content: string;
+  key: string;
+  isOptional?: boolean;
+}
+
+interface TeamContactsProps {
+  disabled?: boolean;
+  hasEA: boolean;
+  userAttributes?: UserAttribute[];
+  secondTechLead: boolean;
+  secondTechLeadOnClick: () => void;
+}
 
 export default function TeamContacts({
   disabled,
+  hasEA = false,
+  userAttributes = [],
   secondTechLead,
   secondTechLeadOnClick,
-}: {
-  disabled?: boolean;
-  secondTechLead: boolean;
-  secondTechLeadOnClick: () => void;
-}) {
+}: TeamContactsProps) {
   const {
     setValue,
     watch,
     formState: { errors },
+    register,
   } = useFormContext();
-  const handleSecondTechLeadClick = async () => {
-    if (secondTechLead) {
-      const res = await openConfirmModal({
-        content: 'Are you sure you want to remove the secondary technical lead from this product?',
-      });
-      if (res?.state.confirmed) {
-        secondTechLeadOnClick();
-      }
-    } else {
+
+  const users: { [key: string]: Partial<User> | null } = {};
+
+  userAttributes.forEach((attr) => {
+    const user = watch(attr.key);
+    users[attr.key] = user || null;
+  });
+
+  const handleUpdateContact = async (index: number) => {
+    const { state } = await openUserPickerModal({});
+    const updatedUser = { ...state.user };
+    const key = userAttributes[index].key;
+    setValue(key, updatedUser, { shouldDirty: true });
+  };
+
+  const handleRemoveSecondaryTechnicalLead = async () => {
+    const key = userAttributes.find((attr) => attr.isOptional)?.key;
+    if (!key) return null;
+
+    const res = await openConfirmModal({
+      content: 'Are you sure you want to remove the secondary technical lead from this product?',
+    });
+    if (res.state.confirmed) {
       secondTechLeadOnClick();
+      setValue(key, null, { shouldDirty: true });
     }
   };
 
-  const phoneNumber = watch('supportPhoneNumber') || '';
+  if (!hasEA) {
+    userAttributes = userAttributes.filter((_, index) => index !== 3);
+  }
 
-  const handlePhoneNumberChange = (value: string) => {
-    const numericValue = value.replace(/[()\s-]/g, '');
-    setValue('supportPhoneNumber', numericValue || null);
-  };
+  const filteredUserAttributes = userAttributes.map(({ role, key, isOptional }, index) => {
+    const user = users[key];
+    const canShow = user && !disabled && isOptional;
+
+    return (
+      <Table.Tr key={index}>
+        <Table.Td>
+          {role}
+          {isOptional && <span className="italic font-bold"> (Optional)</span>}
+        </Table.Td>
+        <Table.Td>
+          <Group
+            gap="sm"
+            onClick={() => handleUpdateContact(index)}
+            className={cn({
+              'cursor-pointer': !disabled,
+            })}
+          >
+            <Avatar src={getUserImageData(user?.image)} size={36} radius="xl" />
+            <div>
+              <Text size="sm" className="font-semibold">
+                {user?.id ? (
+                  <div>
+                    {formatFullName(user)}
+                    {user?.ministry && (
+                      <Badge color="dark" variant="light" className="ml-1">
+                        {user?.ministry}
+                      </Badge>
+                    )}
+                  </div>
+                ) : (
+                  <UnstyledButton className="text-gray-700 hover:underline">Click to select user</UnstyledButton>
+                )}
+              </Text>
+              <Text size="xs" opacity={0.5}>
+                {user?.email}
+              </Text>
+            </div>
+          </Group>
+        </Table.Td>
+
+        <Table.Td>
+          {user?.jobTitle && (
+            <div>
+              <Badge color="primary" variant="light">
+                {user?.jobTitle}
+              </Badge>
+            </div>
+          )}
+          {user?.officeLocation && (
+            <div>
+              <Badge color="info" variant="gradient">
+                {user?.officeLocation}
+              </Badge>
+            </div>
+          )}
+        </Table.Td>
+
+        <Table.Td className="italic">{formatDate(user?.lastSeen) || <span>has not yet logged in</span>}</Table.Td>
+        <Table.Td>
+          {canShow && (
+            <Button color="danger" size="sm" onClick={handleRemoveSecondaryTechnicalLead} leftSection={<IconMinus />}>
+              Remove
+            </Button>
+          )}
+        </Table.Td>
+      </Table.Tr>
+    );
+  });
 
   return (
-    <div className="">
-      <div className="mt-6 2xl:mt-10 grid grid-cols-1 sm:grid-cols-2 gap-x-8 lg:gap-x-24 gap-y-8">
-        <div className="flex flex-col justify-between">
-          <div>
-            <h3 className="text-base 2xl:text-xl font-semibold leading-7 text-gray-900">Product Owner (PO)</h3>
-            <p className="mt-4 text-base leading-6 text-gray-600">
-              Tell us about the Product Owner (PO). This is typically the business owner of the application. We will use
-              this information to contact them with any non-technical questions. Please use only IDIR linked email
-              address below.
-            </p>
-          </div>
-          <AsyncAutocomplete
-            disabled={disabled}
-            name="projectOwner"
-            className="mt-8"
-            label="Product Owner email"
-            placeHolder="Search product owner's IDIR email address"
-          />
-        </div>
-
-        <div className="flex flex-col justify-between">
-          <div>
-            <h3 className="text-base 2xl:text-xl font-semibold leading-7 text-gray-900">Technical Lead (TL)</h3>
-            <p className="mt-4 text-base leading-6 text-gray-600">
-              This is typically the DevOps specialist. We use this information to contact them with technical questions
-              or notify them about platform events. You require a Primary Technical Lead, a Secondary Technical Lead is
-              optional. Please use only IDIR linked email address below. <br />
-              <span className="text-gray-600">
-                Kindly make sure the Product Owner&#40;PO&#41; and the Primary Technical Lead &#40;TL&#41; are{' '}
-                <span className="text-red-600">NOT</span> the same person.
-              </span>
-            </p>
-          </div>
-          <AsyncAutocomplete
-            disabled={disabled}
-            name="primaryTechnicalLead"
-            className="mt-8"
-            label="Technical Lead email"
-            placeHolder="Search technical lead's IDIR email address"
-          />
-        </div>
-        <div className="mt-6 flex flex-col justify-between">
-          {!disabled && <SecondTechLeadButton clicked={secondTechLead} onClick={handleSecondTechLeadClick} />}
-          {secondTechLead ? (
-            <div className="mt-6">
-              <div>
-                <h3 className="text-base 2xl:text-xl font-semibold leading-7 text-gray-900">Technical Lead (TL)</h3>
-                <p className="mt-4 text-base leading-6 text-gray-600">
-                  This is typically the DevOps specialist. We use this information to contact them with technical
-                  questions or notify them about platform events. You require a Primary Technical Lead, a Secondary
-                  Technical Lead is optional. Please use only IDIR linked email address below.
-                </p>
-              </div>
-              <AsyncAutocomplete
-                disabled={disabled}
-                name="secondaryTechnicalLead"
-                className="mt-8"
-                label="Technical Lead email"
-                placeHolder="Search secondary technical lead's IDIR email address"
-              />
-            </div>
-          ) : null}
-        </div>
-        <div className="mt-6 flex flex-col flex-start">
-          <h3 className="text-base 2xl:text-xl font-semibold leading-7 text-gray-900">
-            After-Hours support contact (optional)
-          </h3>
-          <p className="my-4 text-base leading-6 text-gray-600">
-            For Business Mission Critical Applications Only. You can specify a phone number of your team member who
-            should be contacted by the BC Government&apos;s Service Desk (7-7000) if an issue is reported for your
-            product outside of business hours
-          </p>
-          <IMaskInput
-            className="w-fit"
-            mask="+1 (000) 000-0000"
-            unmask={true}
-            value={phoneNumber}
-            onAccept={handlePhoneNumberChange}
-            placeholder="+1 (999) 999-9999"
-          />
-          <p className="mt-3 text-sm leading-6 text-gray-600">{getErrorMessage(errors.supportPhoneNumber)}</p>
-        </div>
-      </div>
-    </div>
+    <>
+      <>
+        <p className="mt-5 ml-2">Here&apos;s an overview of the key contacts for your product:</p>
+        <ul className="list-disc pl-5 ml-2">
+          {userAttributes.map((attr, index) => (
+            <li key={index}>
+              <span className="font-semibold">{attr.role}</span>: {attr.content}
+            </li>
+          ))}
+        </ul>
+      </>
+      <input type="hidden" {...register('primaryTechnicalLead')} />
+      <p className="mt-5 text-red-500">{errors.primaryTechnicalLead?.message?.toString()}</p>
+      <Table.ScrollContainer minWidth={800}>
+        <Table verticalSpacing="sm">
+          <Table.Thead>
+            <Table.Tr>
+              <Table.Th>Title</Table.Th>
+              <Table.Th>User</Table.Th>
+              <Table.Th>Position</Table.Th>
+              <Table.Th>Last active</Table.Th>
+            </Table.Tr>
+          </Table.Thead>
+          <Table.Tbody>{filteredUserAttributes}</Table.Tbody>
+        </Table>
+      </Table.ScrollContainer>
+      <SupportContact />
+    </>
   );
 }
