+++ app/constants/user.ts
@@ -124,3 +124,34 @@ export const userSorts = [
     sortOrder: Prisma.SortOrder.desc,
   },
 ];
+
+export const userAttributes = [
+  {
+    role: 'Project Owner (PO)',
+    content:
+      'This is the business owner of the application, and their contact information will be used for non-technical inquiries.',
+    key: 'projectOwner',
+    isOptional: false,
+  },
+  {
+    role: 'Primary Technical Lead (TL)',
+    content:
+      'This is the DevOps specialist that handles technical queries and platform updates. A Primary Technical Lead is required.',
+    key: 'primaryTechnicalLead',
+    isOptional: false,
+  },
+  {
+    role: 'Secondary Technical Lead (TL)',
+    content:
+      'This is typically the DevOps specialist. We use this information to contact them with technical questions or notify them about platform events. You require a Primary Technical Lead, a Secondary Technical Lead is optional.',
+    key: 'secondaryTechnicalLead',
+    isOptional: true,
+  },
+  {
+    role: 'Expense Authority (EA)',
+    content:
+      'Grants individuals permission to incur organizational expenses within set limits and guidelines. Use only an IDIR-linked email address below.',
+    key: 'expenseAuthority',
+    isOptional: false,
+  },
+];
