+++ app/helpers/mock-users.ts
@@ -1,3 +1,4 @@
+import { User } from '@prisma/client';
 import _compact from 'lodash-es/compact';
 import { Session } from 'next-auth';
 import { generateSession } from '@/core/auth-options';
