+++ app/validation-schemas/shared.ts
@@ -10,7 +10,7 @@ export const RequestDecision = {
 export type RequestDecision = (typeof RequestDecision)[keyof typeof RequestDecision];
 
 export const userSchema = z.object({
-  id: z.string(),
+  id: z.string().max(24),
   firstName: z.string().min(1, 'First name is required').max(50, 'First name must be 50 characters or less'),
   lastName: z.string().min(1, 'Last name is required').max(50, 'Last name must be 50 characters or less'),
   email: z
